third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-zoom-and-background-size-expected.txt

Checks that getComputedStyle() on a zoomed element returns the right thing.
Results while display:block
Property	Pass?	Set Value	Computed Value
background-position	PASS	10px 10px	10px 10px
background-size	PASS	400px 300px	400px 300px
-webkit-border-horizontal-spacing	PASS	10px	10px
-webkit-border-vertical-spacing	PASS	10px	10px
border-top-style	PASS	solid	solid
border-top-width	PASS	2px	2px
border-right-style	PASS	solid	solid
border-right-width	PASS	2px	2px
border-bottom-style	PASS	solid	solid
border-bottom-width	PASS	2px	2px
border-left-style	PASS	solid	solid
border-left-width	PASS	2px	2px
border-top-left-radius	PASS	5px	5px
border-top-right-radius	PASS	5px	5px
border-bottom-left-radius	PASS	5px	5px
border-bottom-right-radius	PASS	5px	5px
outline-style	PASS	solid	solid
outline-width	PASS	2px	2px
-webkit-column-rule-width	PASS	10px	10px
-webkit-column-rule-style	PASS	solid	solid
-webkit-column-width	PASS	80px	80px
-webkit-column-gap	PASS	20px	20px
-webkit-mask-position	PASS	10px 10px	10px 10px
-webkit-mask-size	PASS	10px 10px	10px 10px
-webkit-perspective	PASS	400px	400px
-webkit-perspective-origin	PASS	20px 20px	20px 20px
-webkit-text-stroke-width	PASS	2px	2px
-webkit-transform-origin	PASS	10px 10px	10px 10px
position	PASS	absolute	absolute
left	PASS	20px	20px
top	PASS	20px	20px
right	PASS	50px	50px
bottom	PASS	50px	50px
font-size	PASS	20px	20px
width	PASS	400px	400px
max-width	PASS	900px	900px
min-width	PASS	200px	200px
height	PASS	250px	250px
max-height	PASS	600px	600px
min-height	PASS	200px	200px
letter-spacing	PASS	2px	2px
word-spacing	PASS	10px	10px
margin-top	PASS	10px	10px
margin-right	PASS	10px	10px
margin-bottom	PASS	10px	10px
margin-left	PASS	10px	10px
padding-top	PASS	10px	10px
padding-right	PASS	10px	10px
padding-bottom	PASS	10px	10px
padding-left	PASS	10px	10px
text-indent	PASS	10px	10px
Results while display:none
Hidden Property	Pass?	Set Value	Computed Value
background-position	PASS	10px 10px	10px 10px
background-size	PASS	400px 300px	400px 300px
-webkit-border-horizontal-spacing	PASS	10px	10px
-webkit-border-vertical-spacing	PASS	10px	10px
border-top-style	PASS	solid	solid
border-top-width	PASS	2px	2px
border-right-style	PASS	solid	solid
border-right-width	PASS	2px	2px
border-bottom-style	PASS	solid	solid
border-bottom-width	PASS	2px	2px
border-left-style	PASS	solid	solid
border-left-width	PASS	2px	2px
border-top-left-radius	PASS	5px	5px
border-top-right-radius	PASS	5px	5px
border-bottom-left-radius	PASS	5px	5px
border-bottom-right-radius	PASS	5px	5px
outline-style	PASS	solid	solid
outline-width	PASS	2px	2px
-webkit-column-rule-width	PASS	10px	10px
-webkit-column-rule-style	PASS	solid	solid
-webkit-column-width	PASS	80px	80px
-webkit-column-gap	PASS	20px	20px
-webkit-mask-position	PASS	10px 10px	10px 10px
-webkit-mask-size	PASS	10px 10px	10px 10px
-webkit-perspective	PASS	400px	400px
-webkit-perspective-origin	PASS	20px 20px	20px 20px
-webkit-text-stroke-width	PASS	2px	2px
-webkit-transform-origin	PASS	10px 10px	10px 10px
position	PASS	absolute	absolute
left	PASS	20px	20px
top	PASS	20px	20px
right	PASS	50px	50px
bottom	PASS	50px	50px
font-size	PASS	20px	20px
width	PASS	400px	400px
max-width	PASS	900px	900px
min-width	PASS	200px	200px
height	PASS	250px	250px
max-height	PASS	600px	600px
min-height	PASS	200px	200px
letter-spacing	PASS	2px	2px
word-spacing	PASS	10px	10px
margin-top	PASS	10px	10px
margin-right	PASS	10px	10px
margin-bottom	PASS	10px	10px
margin-left	PASS	10px	10px
padding-top	PASS	10px	10px
padding-right	PASS	10px	10px
padding-bottom	PASS	10px	10px
padding-left	PASS	10px	10px
text-indent	PASS	10px	10px
This div has a zoom value of "2." It has a width of 300px. Its background size is: 400px by 300px.
